tap dance/sticky key thing on bottom right pinky
the bottom right pinky SHOULD send a / when pressed, a ? when tapped after tapping the sticky left thumb, and a ! when pressed while HOLDING the sticky left thumb.

I changed the tap-dance behavior of the sticky thumb key so it sends a left shift when tapped and a right shift when held. Fingers crossed.