New word delete and modified shift behavior
added a word delete to a pinkie combo

changed the shift behavior from a plain sticky thing to a hold-tap that outputs shift if held and sticky shift if tapped, according to the comment here: https://github.com/zmkfirmware/zmk/issues/986#issuecomment-1039736293

This is different from how I tried to implement this last time. I think this will work.