A few small fixes before a bigger, less-confident fix
change RT4 into a layer-tap with ' on the tap and the control layer on the hold so I don't have to change layers to type the apostrophe

change RB2 and RB3 to ; and : so I don't have to hold down the , and . keys. This also frees up these keys for other stuff.

add shift and control to the nav/num layers (I don't remember why I want control, but shift will let me add shift while I'm navigating around, instead of having to leave the layer and then come back for shift.

Fix the Tri Layer so I can use it again. Added a sticky layer key to the control layer